Add files via upload
This is a model to calculate  the demand for shared bikes with the available independent variables. It will be used by the management to understand how exactly the demands vary with different features. They can accordingly manipulate the business strategy to meet the demand levels and meet the customer's expectations. Further, the model will be a good way for management to understand the demand dynamics of a new market.
 This Linear Regression  Model has been made  through 5 important steps:
Step 1: Reading and Understanding the Data
Step 2 : Visualising the data
Step 3 : Preparing the data for Modelling
Step 4 : Splitting the data into Training and Test data
Step 5 : Residual Analysis
Step 6 : Predictions and Evaluations on the Test Data